Remove legacy code
