+++ test/runtests.jl
@@ -5,5 +5,5 @@ using Test
     include("test_domain_utils.jl")
     include("test_treemanipulations.jl")
     include("test_pattern_match.jl")
-    #include("test_propagators.jl")
+    #include("test_pattern_match_edgecases.jl")
 end
